Add core extended_hotkeys file
This has most of our core logic.

The function that you should be interested in is the
'addExtendedCommand' function.

You use this very similarly to <nuke.Menu>.addCommand
The difference is the first argument has to be a <nuke.Menu> instance
and the 'command' arg from the original function is now called
'commands' and is required.
'commands' can be the same data types as the original, str and
types.FunctionType, but it can also be an iterable of the former two
types.

If it is an iterable and you pass more than 1 value, we will implement
timer logic to allow you to have multiclicking.
You have the option to also pass a 'timeout' arg which will only be used
if you pass multiple commands in.

The timeout is a total length of time from your first click until it
executes. Subsequent keypresses will not extend the time, however they
will allow you to iterate through the command list that you passed in.

For example, lets say you have two functions `foo()` and `bar()` and you pass
them both as a list [foo, bar] with a timeout of 5000.
If you press your hotkey once, 5 sec after that, `foo()` will get
executed.
If you press your hotkey once, and then 1 second later press it again,
the timer has not run out yet so internally we record that second press
and `bar()` will be executed.

We have some fast exit logic in place that will detect if we are at the
end of the callback_list. If we are, we will immediatly execute the last
callback. This functionality can be disabled by passing
"fast_exit=False" to 'addExtendedCommand'.